Refactor code for cleaner directories.
