+++ src/main/java/duke/ui/MainWindow.java
@@ -1,6 +1,6 @@
-package duke;
+package duke.ui;
 
-import duke.ui.UserInterface;
+import duke.Duke;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
